+++ packages/augur-ui/src/modules/create-market/components/create-market-view/create-market-view.tsx
@@ -2,18 +2,24 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Helmet } from "react-helmet";
 
-import { LANDING, SCRATCH, TEMPLATE } from "modules/create-market/constants";
+import { Getters } from '@augurproject/sdk';
 
+import { LANDING, SCRATCH, TEMPLATE } from "modules/create-market/constants";
 import Form from "modules/create-market/containers/form";
 import Landing from "modules/create-market/containers/landing";
 import Styles from "modules/create-market/components/create-market-view/create-market-view.styles.less";
+import { Universe } from "@augurproject/sdk/src/state/getter";
+import { CategoryStats } from "@augurproject/sdk/src/state/getter/Markets";
 
 interface CreateMarketViewProps {
+  universe: Universe;
+  getCategoryStats: Function;
 }
 
 interface CreateMarketViewState {
   selected: number;
   page: number;
+  categoryStats: CategoryStats | null;
 }
 
 export default class CreateMarketView extends React.Component<
@@ -30,6 +36,15 @@ export default class CreateMarketView extends React.Component<
 
   componentDidMount() {
     window.scrollTo(0,0);
+      this.props.getCategoryStats(
+        this.props.universe.id,
+        (err, result: Getters.Markets.CategoryStats) => {
+          if (err) return console.log('Error getCategoryStats:', err);
+          this.setState({
+            categoryStats: result,
+          });
+        }
+      );
   }
 
   render() {
@@ -41,9 +56,9 @@ export default class CreateMarketView extends React.Component<
           <title>Create Market</title>
         </Helmet>
         {page === LANDING &&
-          <Landing updatePage={this.updatePage} />
+          <Landing updatePage={this.updatePage} categoryStats={this.state.categoryStats} />
         }
-        {page === TEMPLATE && <Form {...this.props} template updatePage={this.updatePage} />}
+        {page === TEMPLATE && <Form {...this.props} template updatePage={this.updatePage} categoryStats={this.state.categoryStats} />}
         {page === SCRATCH && <Form {...this.props} updatePage={this.updatePage} />}
       </section>
     );
