+++ packages/augur-ui/src/modules/create-market/containers/landing.ts
@@ -4,7 +4,6 @@ import {
   updateNewMarket,
   clearNewMarket
 } from 'modules/markets/actions/update-new-market';
-import { getCategoryStats } from 'modules/create-market/actions/get-category-stats';
 import Landing from 'modules/create-market/landing';
 import getValue from 'utils/get-value';
 
@@ -18,7 +17,6 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch => ({
   updateNewMarket: data => dispatch(updateNewMarket(data)),
   clearNewMarket: () => dispatch(clearNewMarket()),
-  getCategoryStats: (universeId, cb) => dispatch(getCategoryStats(universeId, cb)),
 });
 
 const LandingContainer = withRouter(
